requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
seaborn>=0.12
plotly>=5.15
pykrx>=1.0
FinanceDataReader>=0.9
requests>=2.31
pyyaml>=6.0
tqdm>=4.65
ta>=0.10.2
beautifulsoup4>=4.12
openpyxl>=3.1
jinja2>=3.1
scipy>=1.10
pyarrow>=12.0
pytest>=7.4
scikit-learn>=1.3
hmmlearn>=0.3
lxml>=4.9
